Changes added: 1. User.model added, Movie.model and Celebrity.model modified, 2. routes movies/user and celebrities/user created, 3. partials for the routes: /movies, /movies/user, /celebrities and /celebrities/user created, 4. view of a list of movies created per user and view of a list of celebrities created per user , 5. details page containing user name (created by [user name]), 6. only logged in user can edit and delete the movie and the celebrity

+++ starter-code/routes/celebrities.js
@@ -16,12 +16,29 @@ router.get('/', (req, res, next) => {
 });
 
 
+// GET /celebrities/user
+router.get('/user', (req, res, next) => {
+
+  // only logged in users can get this view
+  // in the worse case (if there would be celebrities in the data bank created not by users but by a developer) when the user would log out and open the /movies/user route, he/she would be able to see movies for user id = 0
+  if (!req.session.user) {
+    res.redirect('/login');
+  } else {
+    Celebrity.find({ createdBy: req.session.user }, null, { sort: { name: 1 } }).then((celebritiesFromDB) => {
+      console.log(celebritiesFromDB);
+
+      res.render('celebrities/user-index', { allTheCelebrities: celebritiesFromDB })
+    })
+  }
+});
+
+
 // GET /celebrities/new
 // show form to the user
 // tricky case: keep it above the code rendering /:id route! in other case browser will try to render "new" as an id and it will cause an error
 router.get('/new', (req, res, next) => {
 
-  // only loggeg in users can see the form
+  // only logged in users can see the form
   if (!req.session.user) {
     res.redirect('/login');
   } else {
@@ -33,25 +50,13 @@ router.get('/new', (req, res, next) => {
 });
 
 
-// GET /celebrities/:id
-router.get('/:id', (req, response, next) => {
-
-  Celebrity.findById(req.params.id)
-    .populate('movies')
-    .then((celebrity) => {
-      response.render('celebrities/show', celebrity);
-    });
-
-});
-
-
 // pick up data from submitted form
 // POST /celebrities/new
 // name: req.body.name = ModelPropertyName.req.body.formInputFieldName
 router.post('/new', (req, res, next) => {
 
   console.log(req.body);
-  // only loggeg in users can create a new celebrity
+  // only logged in users can create a new celebrity
   if (!req.session.user) {
     res.redirect('/login');
   } else {
@@ -71,40 +76,65 @@ router.post('/new', (req, res, next) => {
 });
 
 
+// GET /celebrities/:id
+router.get('/:id', (req, response, next) => {
+
+  Celebrity.findById(req.params.id)
+    .populate('movies')
+    .then((celebrity) => {
+      response.render('celebrities/show', celebrity);
+    });
+
+});
+
 
 // POST /celebrities/id/delete
 router.post('/:id/delete', (req, res, next) => {
 
-  Celebrity.findByIdAndDelete(req.params.id) // builds a query to get and delete a document
-    .exec() // executes the query and returns a promise for the document
-    .then(dbCelebrity => { // takes deleted celebrity object
+  // only logged in users can delete the celebrity
+  if (!req.session.user) {
+    res.redirect('/login');
+  } else {
+    Celebrity.findByIdAndDelete(req.params.id) // builds a query to get and delete a document
+      .exec() // executes the query and returns a promise for the document
+      .then(dbCelebrity => { // takes deleted celebrity object
 
-      return Promise.all(dbCelebrity.movies.map(movieId => Movie.findByIdAndUpdate(movieId, { $pull: { celebrities: dbCelebrity._id } }).exec()));
-      // dbCelebrity.movies.map() > use .map() on a celebrity object, get the movies array (in that object)
-      // findByIdAndUpdate() method builds another query to get and update a document (pull out the celebrity id from the movie document)
+        return Promise.all(dbCelebrity.movies.map(movieId => Movie.findByIdAndUpdate(movieId, { $pull: { celebrities: dbCelebrity._id } }).exec()));
+        // dbCelebrity.movies.map() > use .map() on a celebrity object, get the movies array (in that object)
+        // findByIdAndUpdate() method builds another query to get and update a document (pull out the celebrity id from the movie document)
 
-    }).then(() => {
-      res.redirect('/celebrities');
-    });
-
+      }).then(() => {
+        res.redirect('/celebrities');
+      });
+  }
 });
 
 
 // GET celebrities/id/edit
 router.get('/:id/edit', (req, res, next) => {
 
-  Celebrity.findById(req.params.id).then((celebrity) => {
-    res.render('celebrities/edit', celebrity);
-  });
+  // only logged in users can edit the celebrity
+  if (!req.session.user) {
+    res.redirect('/login');
+  } else {
+    Celebrity.findById(req.params.id).then((celebrity) => {
+      res.render('celebrities/edit', celebrity);
+    });
+  }
 });
 
 
 // POST celebrities/id/edit
 router.post('/:id/edit', (req, res, next) => {
 
-  Celebrity.findByIdAndUpdate(req.params.id, { name: req.body.name, occupation: req.body.occupation, catchPhrase: req.body.catchPhrase }).then(() => {
-    res.redirect('/celebrities');
-  });
+  // only logged in users can edit the celebrity
+  if (!req.session.user) {
+    res.redirect('/login');
+  } else {
+    Celebrity.findByIdAndUpdate(req.params.id, { name: req.body.name, occupation: req.body.occupation, catchPhrase: req.body.catchPhrase }).then(() => {
+      res.redirect('/celebrities');
+    });
+  }
 });
